This is a partially completed project
I wrote what's here before really learning
about git, that's why it's such a big commit.
It's going to be a small, and stupid, chess
engine that will be loaded onto an ATmega328P
micro-controller.

So far, the rules are done and the main
function and makefile are set up for testing.
I still need to get the engine working and
to get UART set up for communicating with a PC
so that you can actually play against it.